fix(settings): vertical preview stands in for letter-stacking
QA finding: the App Bar mock drew names as horizontal text on
left/right edges, implying the real bar does too — it doesn't
(AppBarItemView letter-stacks "S/a/f/…" and hides text a slot
can't fit). Mini slots can't fit a stack, so vertical edges
now show the icon, or the name's initial in Name-only mode.

Splits the geometry/scale block into
AppBarPreviewStrip+Geometry.swift — the addition pushed the
file past the 350-line ceiling.

Refs #293.
